Default Ouro inference to BF16 and remove one-off experiment harnesses (#11)
* Add independent incremental Ouro numerical validation (Q1)


* Record complete Q1 evidence and unresolved BF16 diagnosis


* fix: observe oracle ownership across failed cleanup


* docs: remove intermediate reviews, plans and results


* feat: default Ouro inference to BF16


* refactor: remove one-off experiment harnesses and their tests

Remove M1/Q1 plans, contracts, reporting, validation adapters, fixtures, and harness-only tests. Retain the inference runtime and focused model, engine, attention, and KV correctness tests, with historical evidence linked from the README.


* docs: condense usage and core correctness guidance


---------

+++ vllm_lt/entrypoints/cli.py
@@ -19,7 +19,7 @@ def main():
         "--toy", action="store_true", help="Use a tiny RANDOM model and token ID prompts"
     )
     parser.add_argument("--device", default="cpu")
-    parser.add_argument("--dtype", choices=["float32", "float16", "bfloat16"], default="float32")
+    parser.add_argument("--dtype", choices=["float32", "float16", "bfloat16"], default="bfloat16")
     parser.add_argument("--attention-backend", choices=["torch", "triton"], default="torch")
     parser.add_argument("--mode", choices=["refill", "no_refill"], default="refill")
     parser.add_argument("--max-tokens", type=int, default=16)
